requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
redis==5.0.8
pydantic==2.9.2
prometheus-client==0.21.0
alembic==1.13.2
SQLAlchemy==2.0.36
